Correct the triiger DetId for x, y positions

+++ DataFormats/ForwardDetId/src/HGCalTriggerDetId.cc
@@ -57,7 +57,7 @@ int HGCalTriggerDetId::triggerCellX() const {
   for (auto const& v : vc) {
     x += (3 * (v - N) + 2);
   }
-  return (x / vc.size());
+  return (x / static_cast<int>(vc.size()));
 }
 
 int HGCalTriggerDetId::triggerCellY() const {
@@ -70,7 +70,7 @@ int HGCalTriggerDetId::triggerCellY() const {
   for (unsigned int k = 0; k < uc.size(); ++k) {
     y += (2 * uc[k] - (N + vc[k]));
   }
-  return (y / vc.size());
+  return (y / static_cast<int>(vc.size()));
 }
 
 std::vector<int> HGCalTriggerDetId::cellU() const {
